CORE-4354: Refactor V1 packaging code to prepare for V2
Move all classes inside `net.corda.packaging.util` package into `net.corda.packaging.internal`.

Make `CpiIdentifierImpl`, `SignatureCollector`, `TeeInputStream`, `UncloseableInputStream`, and `ZipTweaker` internal.
