+++ libs/packaging/packaging/src/main/kotlin/net/corda/packaging/internal/TeeInputStream.kt
@@ -1,4 +1,4 @@
-package net.corda.packaging.util
+package net.corda.packaging.internal
 
 import java.io.FilterInputStream
 import java.io.InputStream
@@ -7,7 +7,7 @@ import java.io.OutputStream
 /**
  * [InputStream] that also writes its content to the provided [OutputStream] while reading
  */
-class TeeInputStream(inputStream : InputStream, private val destination : OutputStream): FilterInputStream(inputStream) {
+internal class TeeInputStream(inputStream : InputStream, private val destination : OutputStream): FilterInputStream(inputStream) {
     var written = 0
     override fun read(): Int {
         return super.read().also {
